Rename Durango module to Veracruz Client
