+++ veracruz-client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "durango"
+name = "veracruz-client"
 version = "0.3.0"
 authors = ["The Veracruz Development Team"]
 edition = "2018"
