+++ veracruz-client/src/error.rs
